requirements.txt
anyio 
argon2-cffi 
argon2-cffi-bindings 
autopep8 
bcrypt 
certifi
cffi 
click
colorama 
cryptography 
dataclasses 
fastapi 
flit_core 
greenlet 
h11 
idna 
importlib-metadata 
passlib
pycodestyle 
pycparser 
pydantic 
PyJWT 
PyMySQL 
python-dotenv 
six 
sniffio 
SQLAlchemy 
starlette 
toml 
typing_extensions 
uvicorn 
wincertstore
zipp 
